Fix important bug -> change database structure

+++ inc/tagitem.class.php
@@ -21,7 +21,7 @@ class PluginTagTagItem extends CommonDBRelation {
          $query = "CREATE TABLE IF NOT EXISTS `$table` (
             	`id` INT(11) NOT NULL AUTO_INCREMENT,
             	`plugin_tag_tags_id` INT(11) NOT NULL DEFAULT '0',
-            	`items_id` TINYINT(1) NOT NULL DEFAULT '1',
+            	`items_id` INT(11) NOT NULL DEFAULT '1',
             	`itemtype` VARCHAR(255) NOT NULL DEFAULT '' COLLATE 'utf8_unicode_ci',
             	PRIMARY KEY (`id`),
             	INDEX `name` (`itemtype`)
